Move match trie
